Instead of the end start at the beginning
Before that change when you clicked on a table cell it set the cursor at the end of the cell.
The newline changes the size of the cell that results in the text to be longer than what the users see and the cursor goes out of index and it fails to set the cursor in the cell.